-Fix bugs in AssetManager & replace RawModel by the new Mesh class. -Fix clock wrong type for deltatime. -Improve display class: change clear method parameters to a vec4. - Update engine main render method pipeline. - Refactor Loader class and add tangents ! (Also remove old RawModel class). - Refactor the main.cpp file to use the new rendering pipeline ! - Improve material class by storing the member shader as a pointer (+ add getters & setters). - Refactor the whole Mesh class to fit the new rendering pipeline and the Loader class. - Remove old obj_loader class - Remove old rawModel class - Implement new Renderer::render method. - Add new methods to the scene class : addCamera & addLight. Also change clearColor to a vec4. - Add new unbind method to the Shader class. - Add remove_extension method in utility file. - Remove useless vertex class. - Add new tangent attribute to the basic vertex shader.

+++ PXL/clock.h
@@ -1,5 +1,7 @@
 #pragma once
 
+#include <SDL2/SDL.h>
+
 class Clock
 {
 public:
@@ -8,6 +10,6 @@ public:
 
 	void update();
 
-	int m_lastFrame;
-	int m_deltaTime;
+	Uint32 m_lastFrame;
+	float m_deltaTime;
 };
